feat: anime resolve cache
